requirements.txt
# AI Trading Advancements - Core Dependencies
# Week 1: Foundation Dependencies

# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Machine Learning & AI
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
xgboost>=2.0.0
stable-baselines3>=2.0.0
gymnasium>=0.29.0

# Computer Vision & Image Processing
ultralytics>=8.0.0
Pillow>=10.0.0
opencv-python>=4.8.0

# Financial Data APIs
yfinance>=0.2.18
alpha-vantage>=2.3.1
ccxt>=4.0.0

# Technical Analysis & Charting
ta-lib>=0.4.25
mplfinance>=0.12.0

# Sentiment Analysis & NLP
transformers>=4.30.0
vaderSentiment>=3.3.2
textblob>=0.17.1
tweepy>=4.14.0
praw>=7.7.0

# Genetic Algorithms
deap>=1.4.1

# Database & Backend
sqlalchemy>=2.0.0
asyncpg>=0.28.0
redis>=4.5.0

# Task Queue & Async Processing
celery>=5.3.0
faiss-cpu>=1.7.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.7.4,<3.0.0  # Pin for compatibility (avoid breaking changes)
loguru>=0.7.0
tqdm>=4.65.0
pytz>=2023.3

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Vector / Embedding (Week 3 ChromaDB MVP)
chromadb>=0.5.0
openai>=1.0.0
sentence-transformers>=2.7.0
